Move fetchDatasetDetails to api/datasets
