+++ client/src/api/datasets.ts
@@ -1,5 +1,6 @@
 import type { FetchArgType } from "openapi-typescript-fetch";
 
+import { DatasetDetails } from "@/api";
 import { fetcher } from "@/api/schema";
 import { withPrefix } from "@/utils/redirect";
 
@@ -33,6 +34,15 @@ export async function getDatasets(options: GetDatasetsOptions = {}) {
     }
     const { data } = await datasetsFetcher(params);
     return data;
+}
+
+const getDataset = fetcher.path("/api/datasets/{dataset_id}").method("get").create();
+
+export async function fetchDatasetDetails(params: { id: string }): Promise<DatasetDetails> {
+    const { data } = await getDataset({ dataset_id: params.id, view: "detailed" });
+    // We know that the server will return a DatasetDetails object because of the view parameter
+    // but the type system doesn't, so we have to cast it.
+    return data as unknown as DatasetDetails;
 }
 
 const updateHistoryDataset = fetcher.path("/api/histories/{history_id}/contents/{type}s/{id}").method("put").create();
