+++ client/src/stores/datasetStore.ts
@@ -2,8 +2,7 @@ import { defineStore } from "pinia";
 import Vue, { computed, ref } from "vue";
 
 import { DatasetDetails, DatasetEntry, HistoryContentItemBase } from "@/api";
-
-import { fetchDatasetDetails } from "./services/dataset.service";
+import { fetchDatasetDetails } from "@/api/datasets";
 
 export const useDatasetStore = defineStore("datasetStore", () => {
     const storedDatasets = ref<{ [key: string]: DatasetDetails }>({});
